heatmaps/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==4.2.20
django-environ==0.12.0
frozendict==2.4.6
gunicorn==23.0.0
idna==3.10
multitasking==0.0.11
numpy==2.0.2
packaging==24.2
pandas==2.2.3
peewee==3.17.9
platformdirs==4.3.7
psycopg2==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
whitenoise==6.9.0
yfinance==0.2.55
